Change Offset from uint32_t to size_t
This is to avoid a compile warning. It complains that I am casting it to a const void* of a greater size.

+++ Engine/src/Moon/Core/Renderer/VertexBuffer/VertexBufferElement.h
@@ -11,7 +11,7 @@ namespace Moon {
 		std::string Name;
 		ShaderDataType Type;
 		uint32_t Size;
-		uint32_t Offset;
+		size_t Offset;
 		bool Normalized;
 
 		VertexBufferElement() = default;
